--- a/app/controllers/providers/legal_aid_applications_controller.rb
+++ b/app/controllers/providers/legal_aid_applications_controller.rb
@@ -11,7 +11,6 @@ module Providers
     # POST /provider/applications
     def create
       @legal_aid_application = LegalAidApplication.create(provider: current_provider)
-      @legal_aid_application.create_other_assets_declaration!
       go_forward
     end
   end