--- a/app/views/shared/check_answers/_no_link_items.html.erb
+++ b/app/views/shared/check_answers/_no_link_items.html.erb
@@ -24,7 +24,7 @@
     <% end %>
   <% else %>
     <%= check_answer_link(
-          name: name,
+          name: "#{name}_no_link",
           url: url,
           question: question,
           answer: t('generic.none_declared')