Steps that jumps into a journey by working through a large sequence of previous steps, have been replaced with steps that build the application and associated objects first and then step straight into journey at the appropriate point.

On my system, this PR changed the running time of the `cucumber` command:

Before: real (4 min 16 sec), user (1 min 53 sec)
After: real (2 min 7 sec), user (0 min 56 sec)

So time taken to complete the cucumber tests has been halved.

As the new system uses existing factory traits - there is now some randomness in attribute values. So some feature test that assumed a certain set of attributes are present before the changes made by the specific feature test, have been modified (and some removed). 

The basic assumptions I've used are:

- Feature test should test flow rather than specific behaviour (which should be tested in the request specs). For example, when checking answers, that when clicking on a change link, the user is taken on a flow that returns them to the checking answers page, without concentrating on how each page in that flow behaves.

While modifying these tests I've found a couple of bugs that I've fixed:

- There was a dependency in the `other_assets_declaration` page, that the user had previously visited another page.
- The no link items page created two elements with the same id.